Move HUD code into subsystem folder
